SystemInformation: fix value extraction from /proc/cpuinfo
If the given key (e.g. "cpu") is matched at the beginning of any line the value
was returned regardless if that line would be the correct match (e.g.
"cpu family" is not the correct line). Make sure between the key and the colon
are only spaces and tabs.

Change-Id: I8955add3635a0afb73a9f07f7a0bcf600b7c7015