IOStream: Fix check for compiler 'long long' support
In commit bf7f0ddc (Fix KWSys FundamentalType for Universal Binaries,
2009-11-20) we repurposed CMake variables "KWSYS_SIZEOF_${type}" and
dropped use of them from configured headers.  However, use of one was
accidentally left in "IOStream.hxx.in" to record whether the compiler
supports 'long long'.  Replace it with a better test.

Without this, "@KWSYS_SIZEOF_LONG_LONG@" is replaced with "TRUE" which
the preprocessor may evaluate as "0" or "1" depending on whether a
"TRUE" macro is defined.

Change-Id: I8d3e1d3fc5b80ed440dbb8cebd6231068dfbd491