SystemInformation: query memory size with sysconf() on Solaris
The manpage suggests this is available since Solaris 2.3, which is from
1993. No need to spawn a process then for this information. As an extra
bonus we get information about the available physical memory for free,
which was missing before. On 32 bit processes we will still fall back
to the old method as the values returned by a 32 bit process on a 64
bit host are limited to 4 GiB.

Rename QuerySolarisInfo() to QuerySolarisProcessor() as that is now what
this function is about.

Change-Id: I76fe77a7d746d92d1752119f2c2703652c575339