SystemTools: Move typedef mode_t inside the SystemTools class
Cleans up a smidge of global namespace pollution. Allows using this
header file in the same compilation unit with other header files
that may also contain a mode_t typedef, possibly with a different
underlying type.

Actually encountered in the wild when trying to compile a class that
uses both log4cpp and VTK.

Change-Id: Iba5592778fc21d66946026847537d914e56f332b