SystemTools: Avoid unused function warnings
The WindowsFileExists function was originally created to avoid
duplication by the Cygwin and Windows implementations of FileExists.
Now that we use the ANSI API on Cygwin and the Wide API on Windows, one
of the two overloads is always left unused.  Drop the helper functions
and inline the implementations at their respective (single) call sites.

Change-Id: I4d25f98b8b11f0fdce2f360f5f425e0b159b22e5