Directory: Work around PGI problem with Linux Large File Support
The "/usr/include/dirent.h" header on Linux uses a glibc __REDIRECT
to map readdir to readdir64, but that is defined only when using
a GNU-like compiler.  Otherwise it defines readdir to readdir64
via the C preprocessor, but forgets to define dirent to dirent64.

The fix in commit 2f3c4192 (add support for the Portland Compiler to
CMake, 2007-09-17) does not seem to work on all machines.  Instead try
to map dirent to dirent64 whenever the preprocessor defines readdir to
readdir64 with PGI and glibc.

Change-Id: I6d4fc4cb48e5481f6ac8ed3928c0eb6c6ef0564d