Workaround SolarisStudio bug with libstdc++.
Using iostream and cxxabi.h together causes a compile failure

 .../CC-gcc/include/c++/4.8.2/cxxabi.h", line 131: Error: Only one of a
 set of overloaded functions can be extern "C".

See https://community.oracle.com/thread/3644901 for details.

This only happens if the iostream include is before the cxxabi include.

The available solutions are: re-order the includes in
SystemInformation.cxx, skip the test for KWSYS_CXX_HAS_CXXABI, or cause
it to fail by adding the include.  Do the latter for now.

Change-Id: Ia1a6a6fca12eb9a42f4a4be5252f22fb6e74a318