Glob: Add support for directory listing
Glob now supports listing of directories in recursive mode and disabling
listing of directories in non recursive mode.  In recursive mode when
directory listing is enabled directory symlinks are also listed.  For
backward compatibility there is a separate flag for recursive and
non-recursive globbing mode as this functionality was previously
inconsistent.

Change-Id: I6099c0c568a04caf7c9a36a5d476390371115790