SystemTools: Teach RemoveFile to tolerate missing file
Some use cases may have a race condition such that the file to be
removed disappears before we remove it.  Detect when removal fails due
to the file already missing and tolerate it without failing.

On Windows this requires using DeleteFileW instead of _wunlink because
the latter does not seem to always update errno.  Try to delete before
checking permissions because getting permissions will fail if the file
is missing.

Change-Id: If1922a15d742daca6d252c594284800d60cc1fce