SystemTools: Fix GetPath to not affect existing output vector entries
The GetPath method appends paths to an output vector and also converts
paths to forward slashes.  However, the method should not modify entries
in the output vector that existed prior to the call.  Fix it to only fix
slashes in the entries it added to the vector.  Also add a test case
covering this behavior.

Change-Id: Ie098c3eddbe63ed59f0d393aa6c43ea6c395011d