SystemTools: support hiding the translation map
When determining full paths, CMake would treat the as-called path for
the source and build directories as real paths even if they were
symlinks. This causes problems when determining the relative path
between the source and build directories when they are symlinks.

Take the following setup as an example:

    /mnt/src/project/
        src/
        build/ -> /mnt/builds/project/build

With the translation map and /mnt/src/project/build/ as the build
directory, CMake would use ../src as the relative path to the source
tree from the build tree rather than ../../../src/project/src and
lookups using ../src would fail. If the build directory were instead
used *as* /mnt/builds/project/build, everything would work just
fine.

This patch makes the core logic which does this logic optional so that
it can be experimented with in CMake itself.