ConsoleBuf: Always compile test source for host Windows version
Undefine any `_WIN32_WINNT` value that may be defined on the compiler
command line by clients that are targeting the build at a different
version of Windows than the host.  We must build for the host APIs in
order for the test to be able to configure its console correctly.  Since
the test binary is not installed it does not matter if it is not built
for the client's target Windows version.

Change-Id: Ic3e79550713384c45baf25aa624efbcb5d49d0a4