Add an option to split object and interface libs.
Introduce the KWSYS_SPLIT_OBJECTS_FROM_INTERFACE option.  This will
cause the build to generate a separate interface and object library,
instead of a static or shared library.  This will enable kwsys to be
embedded into another library as objects rather than linked as a
seperate library while still maintaining the propagation of target usage
requirements.