SystemInformation: fix potential off-by-one write
in case readlink uses the complete buffer (1024) it will truncate the
result and return 1024 assigned to ll. the subsequent buf[ll] will then be
out of bounds.

the fix assumes, that a truncated result is as useful as experiencing an
error during readlink, hence falling back to /proc/self/exe.