RegularExpression: New RegularExpressionMatch class
The new RegularExpressionMatch allows to keep multiple match
results of a RegularExpression.

The startp and endp pointers of RegularExpression are replaced
by an instance of RegularExpressionMatch.

There also is new thread safe version of RegularExpression::find()
which allows multiple threads to concurrently use the same
RegularExpression on different input strings with independent
RegularExpressionMatch results.