Fix recent regression in installation of namelink
The check for whether or not to install a namelink depends on
KWSYS_BUILD_SHARED, but this variable wasn't being set until later
in CMakeLists.txt, causing the namelink to not be installed if
KWSYS_BUILD_SHARED is not explicitly set. Fix this by moving the
variable assignment closer to the top of the file.