XL: Restore suppression of infinite loop warning in process test
Our use of the `-qsuppress=` flag to suppress this warning was removed
by commit 51e95ba71a (XL: Fix compatibility with newer clang-based XL on
Linux, 2017-04-28) because XL 13.1.[1-6] for Linux does not support the
flag.  XL 16.1 restores support for the flag.  Restore our use of it for
this version.  Suppress the warning on intermediate versions by blocking
all warnings with `-w`.