Process: On Windows do not open stdin file with write permission
The fix in commit ed7f53e0fa (BUG: When handing the child stdin pipe a
file, allow another process to be writing to the file at the same time,
2006-07-17) added `GENERIC_WRITE` but since we do not actually need to
write to the file, only `FILE_SHARE_WRITE` is enough to allow others to
do so.  Remove the request for write permission.

Issue: cmake/cmake#18359