Configure: Drop Large File Support macro definitions
Historically we have defined macros like

* `_LARGEFILE_SOURCE`
* `_LARGEFILE64_SOURCE`
* `_LARGE_FILES`
* `_FILE_OFFSET_BITS`

in `Configure.h` to activate large file support in system headers.  This
requires consuming projects to include the header at the beginning of
every translation unit in order to get the macros defined consistently.

Instead these macros should be defined by the build system on the
compiler command line.  Drop them from KWSys in favor of that approach.
The macros are not needed as often as they once were anyway.