SystemTools: do not special case a junction in RemoveFile
The fix in commit 9f6cd407b (SystemTools: Fix removing of soft links to
directories on Windows., 2017-12-14) made the infastructure added by
commit de83c4d4b (SystemTools: support deleting junction points,
2016-02-17) unnecessary.  Remove it.

Before this commit, when a file on Windows was deleted, it would check
if a file is a junction and then delete just the junction.

Now just skip the special case of a junction and allow the rest of the
code to delete it correctly. Can handle symlinks and junctions.