Terminal: Support GNU make's TTY notification environment variable
Starting with release 4.0, GNU make provides for collecting recipe
output and printing it all at once after the recipe is complete, as
well as ensuring that only one rule is printing output at a time.
This allows parallel builds without mangled output.

However, it means that programs run by make do not have a TTY for
their stdout or stderr: make is collecting this.

GNU make 4.1 and above will set an environment variable MAKE_TERMOUT
to a non-empty value if make believes that after it's done collecting
stdout, the results will be displayed on a TTY.

This patch teaches KWsys to check that environment variable and if
set, proceed as if output is going to a TTY.