STYLE: Make prototype match definition names
Enforce consistency in large projects, where it often happens that a definition
of function is refactored, changing the parameter names, but its declaration in
header file is not updated. With this check, we can easily find and correct
such inconsistencies, keeping declaration and definition always in sync.

Unnamed parameters are allowed and are not taken into account when comparing
function declarations