SystemInformation: improve CPU counting on non-standard architectures
When /proc/cpuinfo does not match the expected layout the number of logical or
physical CPUs can be reported as 0. Adapt the parsing to catch the layout used
on Sparc, and ensure that the number of logical cores is never 0.

While at it slightly improve the CPU frequency checking so it also matches
Linux Sparc32.