Terminal: Disable color output when CLICOLOR is set to 0
This commit completes the color output control support.
Previously, color output could only be forced ON by setting the environment
variable CLICOLOR_FORCE to 1. With this commit, color output can also be
disabled, by setting the environment variable CLICOLOR to 0.
This can be useful e.g. for processing of build logs when the cmake run
happens within a container (where isatty is true), but the processing
happens on the outside.

See http://bixense.com/clicolors/ for documentation of this convention.