testDynamicLoader: Use LIBDL_SO macro if defined.
Starting with glibc v2.34 there is no external version of libdl.so
anymore.  It has been merged into the glibc libc.so.

Assuming libdl.so will be present will break the test, thus the
LIBDL_SO macro should be used.  If the macro is not defined on
the system, we ensure it will be defined using the previously
hardcoded value.

Signed-off-by: Björn Esser <besser82@fedoraproject.org>