Directory: Replace FileData with methods accepting file index
The internal calls to `SystemTools::FileIs{Directory,Symlink}` added by
commit 7573b0fdf (Directory: Add Is{Directory,Symlink} to FileData,
2022-01-11) pass only the file name, which will be interpreted relative
to the current working directory rather than the `Directory` path.
We need to use the absolute path instead.  Rather than trying to
make this available in the `FileData` structure, provide access to
the information through methods accepting the file index.