SystemTools: Add FileId class and GetFileId function
Add a FileId class, which encapsulates the identifier internally used
by the existing SystemTools::SameFile function. The GetFileId function
retrieves a FileId object for the given file or directory. This makes the
file identifier cacheable to the calling code, so that checking whether
one file is the same as many other candidates can be done faster.